fix: removed call to ``apt-get upgrade``.
Upgrading is a bad habit, we want to be sure of our image, and
depending of when this instruction will be launched, the cached
version will vary.

Anyway, we want debian:jessie at the time of debian:jessie and not
another system.

At last, it takes horrible lot of space. And this overweight is
completely variable. At the time I checked, I gained ~100Mb.